Code clean up for model preference script, test and notebook. Create new stats module for common functions
