+++ src/asgardpy/stats/__init__.py
@@ -13,6 +13,13 @@ from asgardpy.stats.stats import (
     get_goodness_of_fit_stats,
     get_ts_target,
 )
+from asgardpy.stats.utils import (
+    fetch_all_analysis_fit_info,
+    get_model_config_files,
+    tabulate_best_fit_stats,
+    copy_target_config,
+    write_output_config_yaml,
+)
 
 __all__ = [
     "check_model_preference_aic",
@@ -21,4 +28,9 @@ __all__ = [
     "get_chi2_sig_pval",
     "get_goodness_of_fit_stats",
     "get_ts_target",
+    "fetch_all_analysis_fit_info",
+    "get_model_config_files",
+    "tabulate_best_fit_stats",
+    "copy_target_config",
+    "write_output_config_yaml",
 ]
